VS 2022 support
This commit adds the initial support for VS 2022 into the main tree.
This is pretty straight forward with the shared project approach now.
There was a single test failure in folds that I need to track down but
that's about the only issue I hit.

A lot of credit goes to @vivlimmsft in commit
17b98f23cae599851d3d797cd2d6b232ec433c24. That is the work that got
VsVim into VS2022 Preview 1. The work involved convinced me that the
shared project approach was necessary going forward. This PR is
effectively doing that and completing the Preview 1 work.