Move to a flat shared project structure (#2943)
The VS team was able to identify the root problem here. The shared
project system expects to essentially see a flat list of projects where
as VsVim had added indirection through various target files. This is a
bug but it also has a very easy work around on the VsVim side: just
flatten the shared project hierarchy.

Tried this on my local machine with VS 2022 and was able to use the nav
bar just fine.

- https://devdiv.visualstudio.com/DevDiv/_workitems/edit/1358472
- https://github.com/dotnet/project-system/issues/7399