Moving FSharp.Core to be a normal reference
Given that it's taking a while to fix the bug around `--standalone` I
looked into what would happen if we deleted the option from our code
base. The code changes themselves are really not that bad and are pretty
straight forward.

The problem is that our VSIX simply cannot control the version of
FSharp.Core that gets loaded at runtime. In the case the FSharp
extension is not installed into VS then, most likely, the version
referenced in the project will load. But if FSharp does exist in VS then
most likely we get that version which may be very different than what we
built and tested against.

Years ago this was a significant issue in the code base. VS didn't have
a good policy for deploying FSharp.Core and we hit several situations
where VS deployed a _very_ old version of it that broke the VsVim
extension. It was the reason for moving to `--standalone`. It's possible
that these days it's not much of an issue though and we can reasonably
depend on the reference.

Still considering this.

https://github.com/dotnet/fsharp/issues/13345