Unwanted selections don't switch to Visual Mode
I'm not sure if this behavior is ever desirable, but created this
PR anyway so that we can discuss.

I don't see any downsides from doing this. Visual mode works from
key input and mouse selections. We just don't automatically switch
to Visual Mode when Visual Studio decides that it wants to select some
text for us by itself.

Fixes #2966